apply some of the refactoring changes for the upcoming earnings moving
